pyproject.toml
[tool.poetry]
name = "rk-utils"
version = "0.1.0"
description = ""
authors = ["Rasul Kireev"]
packages = [
    { include = "src/" },
]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.1"
python-dotenv = "^0.20.0"
PyYAML = "^6.0"
beautifulsoup4 = "^4.11.2"
typesense = "^0.15.0"
pandas = "^1.5.3"
pymongo = "^4.3.3"
paho-mqtt = "^1.6.1"
html2text = "^2020.1.16"
markdown = "^3.4.4"
anthropic = "^0.45.2"
pygithub = "^2.5.0"
google-genai = "^1.9.0"
httpx = "^0.28.1"
nltk = "^3.9.1"
matplotlib = "^3.10.1"
wordcloud = "^1.9.4"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
pylint = "^2.14.4"
flake8 = "^4.0.1"
mypy = "^0.961"

[tool.poetry.group.dev.dependencies]
pygwalker = "^0.1.3"
pre-commit = "^3.0.4"
jupyterlab = "^3.6.1"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
extra_standard_library = ["pytest", "toml", "click"]

[tool.mypy]
ignore_missing_imports = true

[tool.flake8]
max-line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
